Se crea una vista para ue los asesores puedan ver las ventas que realizaron

+++ routes/web.php
@@ -22,58 +22,44 @@ Route::resource('revenue', 'RevenueController');
 Route::resource('simcard', 'SimcardController');
 
 
-Route::post('/', 'SimcardController@importarSimcard')->name('importar.simcard');
+Route::post('/', 'SimcardController@importarSimcard')->name('importar.simcard')->middleware(['admin']);
 Route::post('simcard', 'SimcardController@store')->name('simcard.store');
 Route::get('simcard', 'SimcardController@index')->name('simcard.index');
-Route::put('simcard/{simcard}', 'SimcardController@update')->name('simcard.update');
-Route::get('simcard/{simcard}/edit', 'SimcardController@edit')->name('simcard.edit');
+Route::put('simcard/{simcard}', 'SimcardController@update')->name('simcard.update')->middleware(['admin']);
+Route::get('simcard/{simcard}/edit', 'SimcardController@edit')->name('simcard.edit')->middleware(['admin']);
 
-Route::get('asesor', 'ExcelController@misVentas')->name('mis.ventas');
-Route::get('usuario', 'HomeController@mostrar')->name('listar.usaurios');
-Route::delete('usuario/{id}', 'HomeController@destroy')->name('user.delete');
-Route::get('/usuario/{usuario}', 'HomeController@edit')->name('user.edit');
-Route::put('usuario/{id}', 'HomeController@update')->name('user.update');
+Route::get('asesor', 'ExcelController@misVentas')->name('mis.ventas')->middleware(['asesor','admin']);
+Route::get('usuario', 'HomeController@mostrar')->name('listar.usaurios')->middleware(['admin']);
+Route::delete('usuario/{id}', 'HomeController@destroy')->name('user.delete')->middleware(['admin']);
+Route::get('/usuario/{usuario}', 'HomeController@edit')->name('user.edit')->middleware(['admin']);
+Route::put('usuario/{id}', 'HomeController@update')->name('user.update')->middleware(['admin']);
 
-Route::get('reporte', 'ReporteController@ventas')->name('reporte.ventas');
+Route::get('reporte', 'ReporteController@ventas')->name('reporte.ventas')->middleware(['admin']);
 
-Route::get('reporte/export/', 'ReporteController@export')->name('export');
-Route::get('reporte', 'ReporteController@index')->name('reporte.index');
-Route::post('reporte', 'ReporteController@store')->name('reporte.store');
-Route::get('reporte/{reporte}/edit', 'ReporteController@edit')->name('reporte.edit');
-Route::delete('reporte/{reporte}', 'ReporteController@destroy')->name('reporte.destroy');
+Route::get('reporte/export/', 'ReporteController@export')->name('export')->middleware(['admin']);
+Route::get('reporte', 'ReporteController@index')->name('reporte.index')->middleware(['admin']);
+Route::post('reporte', 'ReporteController@store')->name('reporte.store')->middleware(['admin']);
+Route::get('reporte/{reporte}/edit', 'ReporteController@edit')->name('reporte.edit')->middleware(['admin']);
+Route::delete('reporte/{reporte}', 'ReporteController@destroy')->name('reporte.destroy')->middleware(['admin']);
 
+Route::get('producto', 'ProductoController@index')->name('producto.index')->middleware(['admin']);
+Route::post('producto', 'ProductoController@store')->name('producto.store')->middleware(['admin']);
+Route::get('producto/{producto}/edit', 'ProductoController@edit')->name('producto.edit')->middleware(['admin']);
+Route::delete('producto/{producto}', 'ProductoController@destroy')->name('producto.destroy')->middleware(['admin']);
+Route::put('producto/{producto}', 'ProductoController@update')->name('producto.update')->middleware(['admin']);
 
-Route::get('producto', 'ProductoController@index')->name('producto.index');
-Route::post('producto', 'ProductoController@store')->name('producto.store');
-Route::get('producto/{producto}/edit', 'ProductoController@edit')->name('producto.edit');
-Route::delete('producto/{producto}', 'ProductoController@destroy')->name('producto.destroy');
-Route::put('producto/{producto}', 'ProductoController@update')->name('producto.update');
-
-Route::get('tipo', 'TipoVentaController@index')->name('tipo.index');
-Route::post('tipo', 'TipoVentaController@store')->name('tipo.store');
-Route::get('tipo/{tipo}/edit', 'TipoVentaController@edit')->name('tipo.edit');
-Route::delete('tipo/{tipo}', 'TipoVentaController@destroy')->name('tipo.destroy');
-Route::put('tipo/{tipo}', 'TipoVentaController@update')->name('tipo.update');
-
-
-Route::get('/', 'ReporteController@ventas')->name('reporte.ventas');
-Route::get('reporte', 'ReporteController@index')->name('reporte.index');
-Route::post('reporte', 'ReporteController@store')->name('reporte.store');
-Route::get('reporte/{reporte}/edit', 'ReporteController@edit')->name('reporte.edit');
-Route::delete('reporte/{reporte}', 'ReporteController@destroy')->name('reporte.destroy');
+Route::get('tipo', 'TipoVentaController@index')->name('tipo.index')->middleware(['admin']);
+Route::post('tipo', 'TipoVentaController@store')->name('tipo.store')->middleware(['admin']);
+Route::get('tipo/{tipo}/edit', 'TipoVentaController@edit')->name('tipo.edit')->middleware(['admin']);
+Route::delete('tipo/{tipo}', 'TipoVentaController@destroy')->name('tipo.destroy')->middleware(['admin']);
+Route::put('tipo/{tipo}', 'TipoVentaController@update')->name('tipo.update')->middleware(['admin']);
 
 
-Route::get('producto', 'ProductoController@index')->name('producto.index');
-Route::post('producto', 'ProductoController@store')->name('producto.store');
-Route::get('producto/{producto}/edit', 'ProductoController@edit')->name('producto.edit');
-Route::delete('producto/{producto}', 'ProductoController@destroy')->name('producto.destroy');
-Route::put('producto/{producto}', 'ProductoController@update')->name('producto.update');
-
-Route::get('tipo', 'TipoVentaController@index')->name('tipo.index');
-Route::post('tipo', 'TipoVentaController@store')->name('tipo.store');
-Route::get('tipo/{tipo}/edit', 'TipoVentaController@edit')->name('tipo.edit');
-Route::delete('tipo/{tipo}', 'TipoVentaController@destroy')->name('tipo.destroy');
-Route::put('tipo/{tipo}', 'TipoVentaController@update')->name('tipo.update');
+Route::get('/', 'ReporteController@ventas')->name('reporte.ventas')->middleware(['admin']);
+Route::get('reporte', 'ReporteController@index')->name('reporte.index')->middleware(['admin']);
+Route::post('reporte', 'ReporteController@store')->name('reporte.store')->middleware(['admin']);
+Route::get('reporte/{reporte}/edit', 'ReporteController@edit')->name('reporte.edit')->middleware(['admin']);
+Route::delete('reporte/{reporte}', 'ReporteController@destroy')->name('reporte.destroy')->middleware(['admin']);
